# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Goodfood 3.0 Users Microservice",
    "version": "1.0.0",
    "title": "Goodfood 3.0 Users Service API"
  },
...
  "paths": {
...
    "/address": {
      "post": {
        "tags": ["address"],
        "summary": "Create address record",
        "operationId": "createAddress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid form submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{id}": {
      "patch": {
        "tags": ["user"],
        "summary": "Update user record",
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user to be updated",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid form submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["user"],
        "summary": "Get user by ID",
        "description": "",
        "operationId": "getUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserComposite"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{id}/subscription": {
      "get": {
        "tags": ["subscription"],
        "summary": "Get a user's subscriptions",
        "operationId": "getUserSubscriptions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user who's subscriptions we're looking for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User subscriptions found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription/{id}/weekday": {
      "patch": {
        "tags": ["subscription"],
        "summary": "Update a SCHEDULED subscription's delivery day",
        "operationId": "updateSubscriptionDeliveryDay",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The subscription ID to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "delivery_day": {
                    "$ref": "#/components/schemas/BaseSubscriptionInput/properties/delivery_day"
                  }
                },
                "required": ["delivery_day"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription weekday updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid request submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription/{id}/plan": {
      "patch": {
        "tags": ["subscription"],
        "summary": "Update a SCHEDULED subscription's plan",
        "operationId": "updateSubscriptionPlan",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The subscription ID to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "old_plan_id": {
                    "type": "number",
                    "description": "GF2 legacy plan ID"
                  }
                },
                "required": ["plan_id", "old_plan_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription plan updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid request submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription/{id}/cancellation": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the subscription"
        }
      ],
      "get": {
        "summary": "Get a list of reasons why this subscription was cancelled",
        "tags": ["subscription"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CancellationSelectionEntity"
                      },
                      {
                        "$ref": "#/components/schemas/CancellationSelectionValue"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "operationId": "getCancellationSelections",
        "description": "Returns the reasons for this subscription's cancellation"
      },
      "post": {
        "summary": "",
        "tags": ["subscription"],
        "operationId": "PostSubscriptionCancellation",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CancellationSelectionEntity"
                      },
                      {
                        "$ref": "#/components/schemas/CancellationSelectionValue"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Subscription not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CancellationSelectionValue"
                }
              }
            }
          },
          "description": "An array of selections over cancellation_reason"
        },
        "description": "Sets the subscription into cancellation state with an array of reason selections in the body."
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "AddressRequestBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "user_id": {
                "type": "string"
              },
              "is_default": {
                "type": "boolean"
              }
            },
            "required": ["user_id", "is_default"]
          },
          {
            "$ref": "#/components/schemas/MigrationAddressRequestBody"
          }
        ]
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string"
          },
          "address_line_1": {
            "type": "string"
          },
          "address_line_2": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province_code": {
            "type": "string"
          },
          "country_code": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          },
          "building_type": {
            "type": "string"
          },
          "special_instructions": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "fsa": {
            "type": "string",
            "example": "K1A"
          }
        },
        "required": ["id"]
      },
...
      "UpdateUserRequestBody": {
        "type": "object",
        "properties": {
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "Last"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firebase_id": {
            "type": "string",
            "example": "e1I4xYb9T0U35Od9nfNmaCjgRoS2",
            "description": "firebase_id comes from the ID generated on Firebase Authentication when the user is registered there. It is a random string of 28 characters with a mix of upper- and lower-case characters and numbers. It must be unique."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "pattern": "/^[0-9]{10}$/g"
          },
          "first_name": {
            "type": "string",
            "example": "First"
          },
          "last_name": {
            "type": "string",
            "example": "McLasty"
          },
          "language": {
            "type": "string",
            "example": "en",
            "enum": ["en", "fr"]
          }
        },
        "required": ["id", "firebase_id", "email", "language"]
      },
      "UserComposite": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "properties": {
              "default_postal_code": {
                "type": "string",
                "example": "H4T1K2"
              },
              "addresses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Address"
                }
              },
              "address_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `addresses` array."
              },
              "subscriptions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Subscription"
                }
              },
              "membership_subscription_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `subscriptions` array."
              },
              "scheduled_subscription_ids": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "deprecated": true,
                "description": "This field is deprecated. Please use the data returned in the `subscriptions` array."
              },
              "carts": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "integer",
                      "example": "113"
                    },
                    "delivery_date": {
                      "type": "string",
                      "format": "date",
                      "example": "2021-09-24T00:00:00.000Z"
                    },
                    "type": {
                      "type": "string",
                      "enum": ["CURATED", "STANDARD"]
                    },
                    "status": {
                      "type": "string",
                      "enum": ["ACTIVE", "CANCELLED", "INCOMPLETE"]
                    },
                    "item_count": {
                      "type": "integer",
                      "example": 3
                    }
                  }
                }
              },
              "preference_tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "number_of_recipes": {
            "type": "integer"
          },
          "number_of_portions": {
            "type": "integer"
          }
        },
        "required": ["id", "number_of_recipes", "number_of_portions"]
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "user_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "card_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "address_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "agreement_id": {
            "type": "integer"
          },
          "state": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLATION",
              "TRIAL",
              "UNPAID",
              "CANCELLED",
              "PAUSED",
              "SUSPENDED"
            ]
          },
          "subscription_type": {
            "type": "string",
            "enum": ["SCHEDULED", "MEMBERSHIP"]
          },
          "product_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "frequency_type": {
            "type": "string",
            "enum": ["DAY", "MONTH", "YEAR"]
          },
          "frequency_value": {
            "type": "integer"
          },
          "delivery_day": {
            "type": "string",
            "enum": ["MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY"]
          },
          "next_cycle": {
            "type": "string",
            "format": "date"
          },
          "send_notification": {
            "type": "boolean"
          },
          "started_at": {
            "type": "string",
            "format": "date"
          },
          "product": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Membership"
              },
              {
                "$ref": "#/components/schemas/Plan"
              }
            ]
          }
        },
        "required": [
          "id",
          "user_id",
          "card_id",
          "address_id",
          "state",
          "subscription_type",
          "product_id",
          "frequency_type",
          "frequency_value",
          "next_cycle",
          "send_notification",
          "started_at"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "error description"
          },
          "meta": {
            "description": "Structured representation of the error",
            "example": {
              "errors": [["The fsa field is required."]]
            }
          },
          "status_code": {
            "type": "number"
          }
        }
      },
...
      "CancellationSelectionEntity": {
        "title": "CancellationSelectionEntity",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "subscription_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
...
  }
}